Changed names and refactored
