+++ src/main/java/com/therejects/cab302groupproject/MainMenuLauncher.java
@@ -7,10 +7,10 @@ import javafx.stage.Stage;
 
 import java.io.IOException;
 
-public class LoginLauncher extends Application {
+public class MainMenuLauncher extends Application {
     @Override
     public void start(Stage stage) throws IOException {
-        FXMLLoader fxmlLoader = new FXMLLoader(LoginLauncher.class.getResource("/com/therejects/cab302groupproject/MainMenu.fxml"));
+        FXMLLoader fxmlLoader = new FXMLLoader(MainMenuLauncher.class.getResource("/com/therejects/cab302groupproject/MainMenu.fxml"));
         Scene scene = new Scene(fxmlLoader.load(), 320, 240);
         stage.setTitle("Hello!");
         stage.setScene(scene);
