+++ src/main/resources/com/therejects/cab302groupproject/login-view.fxml
@@ -5,7 +5,7 @@
 <?import javafx.scene.layout.*?>
 
 <StackPane xmlns="http://javafx.com/javafx/21" xmlns:fx="http://javafx.com/fxml/1"
-           fx:controller="com.therejects.cab302groupproject.model.LoginController"
+           fx:controller="com.therejects.cab302groupproject.controller.LoginController"
            stylesheets="@../../../ui/login.css" styleClass="app-root">
 
     <HBox alignment="CENTER">
